Makes sure that environmental objects are behind characters. Closes #48.
Note that this doesn't completely fix the issue and is just a workaround:
when objects are added, one would need to save+reload (e.g. refresh the
page) to get the stack to work. Since eventually we are moving away from
FabricJS, or upgrading to newest vue-fabric, this workaround is acceptable
for the time being.